requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
docopt==0.6.2
Flask==2.0.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.28.5
greenlet==1.1.2
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mysql-connector-python==8.0.27
numpy==1.21.5
packaging==21.3
pandas==1.3.5
paramiko==2.8.1
Pillow==8.4.0
pipreqs==0.4.11
protobuf==3.19.1
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.28
sshtunnel==0.4.0
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
yarg==0.1.9